refactor: costume, payments 분류
