+++ src/main/java/com/api/stuv/domain/shop/dto/costume/CostumeRandomRequest.java
@@ -1,4 +1,4 @@
-package com.api.stuv.domain.shop.dto;
+package com.api.stuv.domain.shop.dto.costume;
 
 import jakarta.validation.constraints.NotNull;
