+++ src/main/java/com/api/stuv/domain/shop/dto/payments/PaymentRequest.java
@@ -1,4 +1,4 @@
-package com.api.stuv.domain.shop.dto;
+package com.api.stuv.domain.shop.dto.payments;
 
 import jakarta.validation.constraints.NotNull;
